[starts #164055166] refactored my code and separated configuration to environment variables
