+++ .env
@@ -1,0 +1,7 @@
+source venv/bin/activate
+
+export FLASK_RUN=run.py
+export DATABASE_URL='postgresql://erycoking:@127.0.0.1/politico'
+export SECRET_KEY='QeVYWfId8LODTNgMRl4-hnAdBYdhuZy_7o9RlfaUKhWdl2G5xEVXfYaEEbJ8QWnNeiDGgQn6wjGmscU7wAVhrA'
+export TEST_DATABASE_URL='postgresql://erycoking:@127.0.0.1/test_politico'
+export FLASK_ENV=development
