+++ politico/api/v2/office/model.py
@@ -1,34 +1,31 @@
 import psycopg2
-from politico.api.v2.db.db import DB
+from politico.api.v2.db import DB
 
-class OfficeTable:
+class OfficeTable(DB):
     """office table"""
 
-    def __init__(self):
-        self.db = DB()
-
     def get_one_office(self, id):
-        office = self.db.fetch_one('office', 'id', id)
+        office = self.fetch_one('office', 'id', id)
         if office is not None:
             return self.office_data(office)
         return None
 
     def get_one_office_by_name(self, name):
-        office = self.db.fetch_one_using_string('office', 'name', name)
+        office = self.fetch_one_using_string('office', 'name', name)
         if office:
             return self.office_data(office)
         return None
 
     def get_offices(self):
         offices = []
-        stored_offices = self.db.fetch_all('office')
+        stored_offices = self.fetch_all('office')
         for office in stored_offices:
             offices.append(self.office_data(office))
         return offices
 
     def create_office(self, office_data):
 
-        conn = self.db.connection()
+        conn = self.connection()
         try:
             cursor = conn.cursor()
             cursor.execute(
@@ -49,7 +46,7 @@ class OfficeTable:
         return None
 
     def update_office(self, id, office_data):
-        conn =  self.db.connection()
+        conn =  self.connection()
         try:
             cursor = conn.cursor()
             cursor.execute(
@@ -70,7 +67,7 @@ class OfficeTable:
         return None
 
     def delete_office(self, id):
-        return self.db.delete_one('office', 'id', id)
+        return self.delete_one('office', 'id', id)
 
     def office_data(self, office):
         office_data = {}
@@ -86,11 +83,11 @@ class OfficeTable:
         # get all candidates in that office
         # get total votes for that votes candidate
         # return the results
-        office = self.db.fetch_one('office', 'id', id)
+        office = self.fetch_one('office', 'id', id)
 
-        candidates = self.db.fetch_all_using_int_key('candidates', 'office', office[0])
+        candidates = self.fetch_all_using_int_key('candidates', 'office', office[0])
         for cand in candidates:
-            votes = self.db.fetch_all_using_int_key('vote', 'candidate', cand[0])
+            votes = self.fetch_all_using_int_key('vote', 'candidate', cand[0])
             total_vote_for_specific_candidate = len(votes)
 
             candidate_result = {
