+++ politico/api/v2/party/routes.py
@@ -123,9 +123,9 @@ def delete_party(current_user, id):
             }), 200)
         else:
             return make_response(jsonify({
-                'status': 500, 
-                'error' : 'Could not delete party with id:{}'.format(id)
-            }), 500)
+                'status': 400, 
+                'error' : 'update or delete on table "party" violates foreign key constraint.\nKey (id)=({}) is referenced on another table'.format(id)
+            }), 400)
