+++ politico/api/v2/petition/model.py
@@ -1,41 +1,38 @@
 import psycopg2
-from politico.api.v2.db.db import DB
+from politico.api.v2.db import DB
 import datetime
 
-class PetitionTable:
+class PetitionTable(DB):
     """petition table"""
 
-    def __init__(self):
-        self.db = DB()
-
     def get_one_petition(self, id):
-        petition = self.db.fetch_one('petition', 'id', id)
+        petition = self.fetch_one('petition', 'id', id)
         if petition is not None:
             return self.petition_data(petition)
         return None
 
     def get_one_petition_by_created_by_and_office(self, created_by, office):
-        petition = self.db.fetch_one_using_two_values('petition', 'created_by', created_by, 'office', office)
+        petition = self.fetch_one_using_two_values('petition', 'created_by', created_by, 'office', office)
         if petition is not None:
             return self.petition_data(petition)
         return None
 
     def get_one_petition_by_created_by(self, created_by):
-        petition = self.db.fetch_one('petition', 'created_by', created_by)
+        petition = self.fetch_one('petition', 'created_by', created_by)
         if petition is not None:
             return self.petition_data(petition)
         return None
 
     def get_petitions(self):
         petitions = []
-        stored_petitions = self.db.fetch_all('petition')
+        stored_petitions = self.fetch_all('petition')
         for petition in stored_petitions:
             petitions.append(self.petition_data(petition))
         return petitions
 
     def create_petition(self, petition_data):
 
-        conn = self.db.connection()
+        conn = self.connection()
         try:
             cursor = conn.cursor()
             cursor.execute(
@@ -58,7 +55,7 @@ class PetitionTable:
         return None
 
     def update_petition(self, id, petition_data):
-        conn =  self.db.connection()
+        conn =  self.connection()
         try:
             cursor = conn.cursor()
             cursor.execute(
@@ -82,7 +79,7 @@ class PetitionTable:
         return None
 
     def delete_petition(self, id):
-        return self.db.delete_one('petition', 'id', id)
+        return self.delete_one('petition', 'id', id)
 
     def petition_data(self, petition):
         petition_data = {}
