+++ politico/api/v2/petition/routes.py
@@ -116,9 +116,9 @@ def delete_petition(current_user, id):
             }), 200)
         else:
             return make_response(jsonify({
-                'status': 500, 
-                'error' : 'Could not delete petition with id:{}'.format(id)
-            }), 500)
+                'status': 400, 
+                'error' : 'update or delete on table "petition" violates foreign key constraint.\nKey (id)=({}) is referenced on another table'.format(id)
+            }), 400)
