+++ politico/api/v2/vote/routes.py
@@ -118,9 +118,9 @@ def delete_vote(current_user, id):
             }), 200)
         else:
             return make_response(jsonify({
-                'status': 500, 
-                'error' : 'Could not delete vote with id:{}'.format(id)
-            }), 500)
+                'status': 400, 
+                'error' : 'update or delete on table "vote" violates foreign key constraint.\nKey (id)=({}) is referenced on another table'.format(id)
+            }), 400)
