+++ politico/test/v1/user/functional/test_user_routes.py
@@ -1,23 +1,7 @@
-""" A test for testing user routes"""
-
-"""imports core modules """
 import pytest
-
-"""import custom modules"""
-from .....config import create_app
+from ....base import test_client
 
 prefix = '/api/v1'
-
-@pytest.fixture(scope = 'module')
-def test_client():
-    """A fixture for creating a test_client"""
-    test_client = create_app().test_client()
-    cxt = create_app().app_context()
-    cxt.push()
-
-    yield test_client
-
-    cxt.pop()
 
 
 def add_user(test_client):
@@ -29,11 +13,12 @@ def add_user(test_client):
         "lastname": "rozay",
         "othername": "king",
         "passport_url": "http://file.com/passports/passport1.png",
-        "phone_number": "0712340000", 
-        "username":"baller1",
-        "password":"password"
+        "phone_number": "0712340000",
+        "username": "baller1",
+        "password": "password"
     }
     return test_client.post(prefix+'/users', json=user_data)
+
 
 def update_user(test_client):
     user_data = {
@@ -45,70 +30,69 @@ def update_user(test_client):
         "othername": "king",
         "passport_url": "http://file.com/passports/passport1.png",
         "phone_number": "0712345679",
-        "username":"baller675",
-        "password":"password"
+        "username": "baller675",
+        "password": "password"
     }
     return test_client.patch(prefix+'/users/1', json=user_data)
 
 
 def test_add_user(test_client):
     """A test for adding a user"""
-    with test_client as c:
-        response = add_user(c)
-        assert response.status_code == 201
-        data  = response.get_json()
-        assert 'data' in data
-        user = data.get('data')[0]
-        print(user)
-        assert 'id' in user and user['id'] == 1
-        assert 'firstname' in user and user['firstname'] == 'king'
+    response = add_user(test_client)
+    assert response.status_code == 201
+    data = response.get_json()
+    assert 'data' in data
+    user = data.get('data')[0]
+    print(user)
+    assert 'id' in user and user['id'] == 1
+    assert 'firstname' in user and user['firstname'] == 'king'
+
 
 def test_get_single_user(test_client):
     """A test for getting a single user"""
-    with test_client as c:
-        add_user(c)
-        response = c.get(prefix+'/users/1')
-        assert response.status_code == 200
-        data  = response.get_json()
-        print(data)
-        assert 'data' in data
-        user = data.get('data')[0]
-        assert 'id' in user and user['id'] == 1
-        assert 'firstname' in user and user['firstname'] == 'king'
+    add_user(test_client)
+    response = test_client.get(prefix+'/users/1')
+    assert response.status_code == 200
+    data = response.get_json()
+    print(data)
+    assert 'data' in data
+    user = data.get('data')[0]
+    assert 'id' in user and user['id'] == 1
+    assert 'firstname' in user and user['firstname'] == 'king'
+
 
 def test_update_user(test_client):
     """A test for getting a single user"""
-    with test_client as c:
-        response = update_user(c)
-        assert response.status_code == 200
-        data  = response.get_json()
-        print(data)
-        assert 'data' in data
-        user = data.get('data')[0]
-        assert 'id' in user and user['id'] == 1
-        assert 'firstname' in user and user['firstname'] == 'bigfish'
+    response = update_user(test_client)
+    assert response.status_code == 200
+    data = response.get_json()
+    print(data)
+    assert 'data' in data
+    user = data.get('data')[0]
+    assert 'id' in user and user['id'] == 1
+    assert 'firstname' in user and user['firstname'] == 'bigfish'
+
 
 def test_getting_all_users(test_client):
     """A test for getting all user"""
-    with test_client as c:
-        add_user(c)
-        response = c.get(prefix+'/users')
-        assert response.status_code == 200
-        data  = response.get_json()
-        assert 'data' in data
-        users = data.get('data')
-        first_user = users.get('1')
-        print(users)
-        assert 'id' in first_user and first_user['id'] == 1
-        assert 'firstname' in first_user and first_user['firstname'] == 'king'
+    add_user(test_client)
+    response = test_client.get(prefix+'/users')
+    assert response.status_code == 200
+    data = response.get_json()
+    assert 'data' in data
+    users = data.get('data')
+    first_user = users.get('1')
+    print(users)
+    assert 'id' in first_user and first_user['id'] == 1
+    assert 'firstname' in first_user and first_user['firstname'] == 'king'
+
 
 def test_delete_user(test_client):
     """A test for deleting a single user"""
-    with test_client as c:
-        add_user(c)
-        response = c.delete(prefix+'/users/1')
-        assert response.status_code == 200
-        data  = response.get_json()
-        print(data)
-        assert 'data' in data
-        assert 'message' in data['data'] and data['data']['message'] == 'user successfully deleted'
+    add_user(test_client)
+    response = test_client.delete(prefix+'/users/1')
+    assert response.status_code == 200
+    data = response.get_json()
+    print(data)
+    assert 'data' in data
+    assert 'message' in data['data'] and data['data']['message'] == 'user successfully deleted'
