+++ requirements.txt
@@ -8,7 +8,7 @@ chardet==3.0.4
 Click==7.0
 coverage==4.5.2
 Flask==1.0.2
-gunicorn==19.7.1
+gunicorn==19.9.0
 httplib2==0.12.0
 idna==2.8
 isort==4.3.4
@@ -17,7 +17,9 @@ Jinja2==2.10
 lazy-object-proxy==1.3.1
 MarkupSafe==1.1.0
 mccabe==0.6.1
+mock==2.0.0
 more-itertools==5.0.0
+pbr==5.1.2
 pluggy==0.8.1
 psycopg2-binary==2.7.7
 py==1.7.0
@@ -26,6 +28,7 @@ PyJWT==1.4.2
 pylint==2.2.2
 pytest==4.2.0
 pytest-cov==2.6.1
+pytest-mock==1.10.1
 python-coveralls==2.9.1
 PyYAML==3.13
 requests==2.21.0
