+++ run.py
@@ -1,8 +1,10 @@
-"""Entry to the application"""
-from politico.config import create_app
+from politico import create_app
+import os
 
+"""Entry to the application"""
 """Initializing application"""
-app = create_app()
+app = create_app(os.getenv('FLASK_ENV'))
 
 """Setting debug to true"""
-app.run(debug=True)
+if __name__ == '__main__':
+    app.run()
